worker_application_latest/requirements.txt
fastapi
boto3
uvicorn
python-dotenv
requests
celery
redis